examples/requirements-google.txt
pyaudio
keyboard
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
